requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
aws_cdk.aws_apigatewayv2_alpha>=2.0.0